pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "deepseek>=1.0.0",
    "langchain>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langgraph>=0.6.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
